refactor: move domain gate to its own pkg
